quant-engine/dnt_quant_lab/backend/requirements.txt
fastapi
uvicorn
pandas
numpy
plotly
requests
google-generativeai
python-dotenv
vnstock3
requests-cache
scikit-learn
scipy
backtrader==1.9.78.123
pandas-ta-classic
supabase